From-SVN: r221365

+++ gcc/testsuite/gcc.target/i386/pr45685.c
@@ -1,8 +1,8 @@
-/* { dg-do compile } */
-/* { dg-options "-O3 -fdump-tree-phiopt1-details" } */
+/* { dg-do compile { target { ! { ia32 } } } } */
+/* { dg-options "-O3" } */
 
-typedef unsigned long int uint64_t;
-typedef long int int64_t;
+typedef unsigned long long int uint64_t;
+typedef long long int int64_t;
 int summation_helper_1(int64_t* products, uint64_t count)
 {
 	int s = 0;
@@ -36,6 +36,4 @@ int summation_helper_2(int64_t* products, uint64_t count)
 	return s;
 }
 
-/* { dg-final { scan-tree-dump-times "converted to straightline code" 2 "phiopt1" } } */
-/* { dg-final { cleanup-tree-dump "phiopt1" } } */
-
+/* { dg-final { scan-assembler-times "cmov" 4 } } */
